Move send message tests to new file
